CME-56: fix smoke test issues
