+++ build.gradle
@@ -13,7 +13,7 @@ plugins {
   id 'au.com.dius.pact' version '4.6.16' //do not change this version or serenity reports will fail to generate
   id 'org.flywaydb.flyway' version '9.22.3'
   id 'org.owasp.dependencycheck' version '12.0.0'
-  id 'net.serenity-bdd.serenity-gradle-plugin' version '4.2.13'
+  id 'net.serenity-bdd.serenity-gradle-plugin' version '4.2.16'
 }
 
 ext['snakeyaml.version'] = '2.0'
@@ -28,7 +28,7 @@ def versions = [
   lombok          : '1.18.36',
   reformLogging   : '6.1.7',
   reformS2sClient : '5.3.0',
-  serenity        : '4.2.13',
+  serenity        : '4.2.15',
   pact_version    : '4.1.7',
   rest_assured    : '5.5.0',
   cucumber        : '7.20.1',
@@ -185,6 +185,8 @@ task integration(type: Test, description: 'Runs the integration tests.', group: 
 check.dependsOn integration
 
 task smoke(type: Test, description: 'Runs the smoke tests.', group: 'Verification') {
+  description = 'Runs Smoke Tests'
+  useJUnitPlatform()
   testClassesDirs = sourceSets.smokeTest.output.classesDirs
   classpath = sourceSets.smokeTest.runtimeClasspath
   finalizedBy aggregate
